Fix setup for development with ganache
The docker compose files were outdated.

Last month apparently the web3 dependency was updated to v5.28 which makes some changes to maxPriorityFeePerGas which were causing the used ganache-cli version to no longer work.
Web3 was making a eth_maxPriorityFeePerGas call that was failing against ganache v6 and thus we were no longer able to install the contracts.

The dockerhub image of ganache is no longer called ganache-cli, but ganache. By switching to trufflesuite/ganache we are installing ganache v7 where maxPriorityFeePerGas is available.

Also I was trying to build on a M1 but the built images were amd64. I updated the readme to reflect that docker can build multi-architecture images by using the buildx command.